creat no item view

+++ ToDoList-SwiftUI/ToDoList-SwiftUI/View/ListView.swift
@@ -12,19 +12,25 @@ struct ListView: View {
     @State private var vm = ListViewModel()
     
     var body: some View {
-        List {
-            ForEach(vm.items){ item in
-                ListRowView(title: item.title ?? "", done: item.done)
-                    .onTapGesture {
-                        item.done.toggle()
-                        vm.updateDone(item: item, done: item.done)
+        ZStack{
+            if vm.items.isEmpty{
+                NoItemView(vm: vm)
+            }else{
+                List {
+                    ForEach(vm.items){ item in
+                        ListRowView(title: item.title ?? "", done: item.done)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    item.done.toggle()
+                                    vm.updateDone(item: item, done: item.done)
+                                }
+                            }
                     }
+                    .onDelete(perform: vm.deleteItem)
+                }
+                .listStyle(.plain)
             }
-            .onDelete(perform: vm.deleteItem)
-            
-            
         }
-        .listStyle(.plain)
         .navigationTitle("ToDo List📝")
         .navigationBarItems(leading: EditButton(),
                             trailing: NavigationLink("Add", destination: {AddView(vm: vm)}))
